+++ InneractiveAdSDK/InneractiveAdSDK.framework/Versions/A/Headers/InneractiveAdSDK-AdConfig.h
@@ -40,23 +40,24 @@ typedef NS_ENUM(NSUInteger, IaAdType) {
 
 /**
  *  @typedef InneractiveVideoFullscreenOrientationMode
+ *
  *  @brief Fullscreen videos interface orientation modes.
+ *
+ *  The default is 'InneractiveVideoFullscreenOrientationModeMaskAll'.
  */
 typedef NS_ENUM(NSInteger, InneractiveVideoFullscreenOrientationMode) {
     /**
-     *  @brief This is default mode.
-     *
-     * A video will be presented in a landscape mode. Can be rotated to any supported landscape orientation.
-     */
-    InneractiveVideoFullscreenOrientationModeForceLandscape = 0,
-    /**
-     *  @brief A video will be presented in portrait.
-     */
-    InneractiveVideoFullscreenOrientationModeForcePortrait = 1,
-    /**
      *  @brief A video will be presented in same orientation as presenting view controller; can be rotated to any supported orientation.
      */
-    InneractiveVideoFullscreenOrientationModeMaskAll = 2
+    InneractiveVideoFullscreenOrientationModeMaskAll = 0,
+    /**
+     * A video will be presented in a landscape mode. Can be rotated to any supported landscape orientation.
+     */
+    InneractiveVideoFullscreenOrientationModeForceLandscape = 1,
+    /**
+     *  @brief A video will be presented in portrait mode.
+     */
+    InneractiveVideoFullscreenOrientationModeForcePortrait = 2,
 };
 
 /**
@@ -95,7 +96,7 @@ typedef NS_ENUM(NSUInteger, IaNativeAdContentType) {
  *
  * An ad spot is a defined placement in your application set aside for advertising.
  */
-@property (nonatomic, copy) NSString *appId;
+@property (nonnull, nonatomic, strong) NSString *appId;
 
 /**
  *  @brief Ad type definition.
@@ -113,13 +114,13 @@ typedef NS_ENUM(NSUInteger, IaNativeAdContentType) {
 /**
  *  @brief A string representing a set of keywords, that should be passed to Inneractive's Ad Server in order to receive more relevant advertising.
  */
-@property (nonatomic, strong) NSString *adKeyWords;
+@property (nullable, nonatomic, strong) NSString *adKeyWords;
 
 /**
  *  @brief User gender. Use for better ad targeting.
  *  @discussion Available values: "M" / "m" / "F" / "f" / "Male" / "Female".
  */
-@property (nonatomic, strong) NSString *userGender;
+@property (nullable, nonatomic, strong) NSString *userGender;
 
 /**
  *  @brief User age. Use for better ad targeting.
@@ -131,7 +132,7 @@ typedef NS_ENUM(NSUInteger, IaNativeAdContentType) {
  *  @brief ZIP Code. Use for better ad targeting.
  *  @discussion Should be specified as NSString instance with standard ZIP code value. E.g. "90210".
  */
-@property (nonatomic, strong) NSString *zipCode;
+@property (nullable, nonatomic, strong) NSString *zipCode;
 
 /**
  *  @brief Current location. Use for better ad targeting.
@@ -141,7 +142,7 @@ typedef NS_ENUM(NSUInteger, IaNativeAdContentType) {
  *
  * If 'shouldAutomaticallyGetLocationData' is enabled, do not use current property, it will be updated automatically by the SDK.
  */
-@property (nonatomic, copy) CLLocation *location;
+@property (nullable, nonatomic, copy) CLLocation *location;
 
 /**
  *  @brief Use to enable a location manager managed by the InneractiveAdSDK.
@@ -161,7 +162,7 @@ typedef NS_ENUM(NSUInteger, IaNativeAdContentType) {
 /**
  *  @brief Video Layout.
  */
-@property (nonatomic, strong, readonly) IaVideoLayout *videoLayout;
+@property (nonnull, nonatomic, strong, readonly) IaVideoLayout *videoLayout;
 
 #pragma mark - Native Ad parameters
 
@@ -192,7 +193,7 @@ typedef NS_ENUM(NSUInteger, IaNativeAdContentType) {
  *
  * 'IaNativeAdAssetPriorityRequired' - Asset (UI) is implemeted by a publisher and is required by a publisher.
  */
-@property (nonatomic, strong, readonly) IaNativeAdAssetDescription *nativeAdAssetsDescription;
+@property (nonnull, nonatomic, strong, readonly) IaNativeAdAssetDescription *nativeAdAssetsDescription;
 
 /**
  *  @brief Native Ad Main Asset content type.
@@ -228,10 +229,34 @@ typedef NS_ENUM(NSUInteger, IaVideoLayoutControlType) {
 };
 
 /**
+ *  @typedef IaVideoClickActionType
+ *  @brief IaVideoClickActionType
+ */
+typedef NS_ENUM(NSUInteger, IaVideoClickActionType) {
+    IaVideoClickActionTypeFullscreen = 0,
+    IaVideoClickActionTypeLandingPage
+};
+
+/**
  *  @class IaVideoLayout
  *  @brief Video controls layout configuration.
  */
 @interface IaVideoLayout : NSObject
+
+/**
+ *  @brief Defines whether the action button (aka: CTA / "Visit Us" / VAST clickthrough) is visible in feed (native and non-fullscreen mode).
+ *
+ *  @discussion This is not OpenRTB Native 1.0 CTA, but VAST protocol CTA. There is also OpenRTB CTA asset in native ad unit (in case a response includes it).
+ *
+ * Default: <b>enabled</b>.
+ * 
+ * If disabled, <b>will not be visible</b> in feed.
+ *
+ * If enabled <b>and</b> CTA native asset is implemented in UI, will be <b>hidden</b> in feed, in order to prevent from showing both CTAs (VAST and OpenRTB).
+ *
+ * If enabled <b>and</b> CTA native asset is NOT implemented in UI, will be <b>visible</b>.
+ */
+@property (nonatomic, getter=isActionButtonVisibleInFeed) BOOL actionButtonIsVisibleInFeed;
 
 /**
  *  @brief Defines whether progress bar is visible in feed (native and non-fullscreen mode).
@@ -241,12 +266,45 @@ typedef NS_ENUM(NSUInteger, IaVideoLayoutControlType) {
 /**
  *  @brief Progress bar progress track fill color.
  */
-@property (nonatomic, strong) UIColor *progressBarFillColor;
+@property (nonnull, nonatomic, strong) UIColor *progressBarFillColor;
 
 /**
  *  @brief Progress bar track color.
  */
-@property (nonatomic, strong) UIColor *progressBarBackColor;
+@property (nonnull, nonatomic, strong) UIColor *progressBarBackColor;
+
+/**
+ *  @brief Defines whether controls are aligned to an actual video rect or to a video player superview (placeholder).
+ *
+ *  @discussion NO by default;
+ */
+@property (nonatomic) BOOL controlsInsideVideoRect;
+
+/**
+ *  @brief Defines click (inside video area) action, while a video has not finished playing.
+ *
+ *  @discussion The default action is to open fullscreen.
+ */
+@property (nonatomic) IaVideoClickActionType videoClickActionType;
+
+/**
+ *  @brief Defines color theme.
+ *
+ *  @discussion Tints text color of text based controls.
+ * Tints image colour of image based controls.
+ *
+ * The default is nil.
+ */
+@property (nullable, nonatomic, strong) UIColor *themeColor;
+
+/**
+ *  @brief Defines background color theme.
+ *
+ *  @discussion Tints background color of text based controls.
+ *
+ * The default is nil.
+ */
+@property (nullable, nonatomic, strong) UIColor *backgroundThemeColor;
 
 /**
  *  @brief Defines controls placement.
@@ -262,6 +320,28 @@ typedef NS_ENUM(NSUInteger, IaVideoLayoutControlType) {
           topRightControlType:(IaVideoLayoutControlType)topRightControlType
         bottomLeftControlType:(IaVideoLayoutControlType)bottomLeftControlType
        bottomRightControlType:(IaVideoLayoutControlType)bottomRightControlType;
+
+/**
+ *  @brief Use to get UI control by control type.
+ *
+ *  @discussion This method should be invoked only after 'InneractiveAdLoaded:' event has been received.
+ * This method should be used in order to customise font, text color, etc.
+ *
+ *  @param type Control type.
+ *
+ *  @return (UIButton *) OR (UILabel *) OR other UIView subclass instance.
+ */
+- (nullable UIView *)controlByType:(IaVideoLayoutControlType)type;
+
+/**
+ *  @brief Use to get 'skip in ...' label of interstitial ad type.
+ *
+ *  @discussion This method should be invoked only after 'InneractiveAdLoaded:' event has been received.
+ * This method should be used in order to customise font, text color, etc.
+ *
+ *  @return Skip label as (UILabel *).
+ */
+- (nullable UILabel *)interstitialSkipLabel;
 
 @end