+++ InneractiveAdSDK/InneractiveAdSDK.framework/Versions/A/Headers/InneractiveAdSDK-NativeAd.h
@@ -9,6 +9,8 @@
 #import "InneractiveAdSDK-Ad.h"
 #import "InneractiveAdSDK-AdConfig.h"
 
+@class IaNativeAdResponseModel;
+
 #pragma mark - Native Ad Interface
 
 /**
@@ -33,17 +35,22 @@
  *
  *  @return IaNativeAd instance.
  */
-- (instancetype)initWithAppId:(NSString *)appId adType:(IaAdType)adType delegate:(id<InneractiveAdDelegate>)adDelegate;
+- (nonnull instancetype)initWithAppId:(nonnull NSString *)appId adType:(IaAdType)adType delegate:(nonnull id<InneractiveAdDelegate>)adDelegate;
 
-- (instancetype)init __attribute__((unavailable("please use 'initWithAppId:adType:delegate:' instead")));
-- (instancetype)initWithFrame:(CGRect)frame __attribute__((unavailable("please use 'initWithAppId:adType:delegate:' instead; the instance of this class should not be used as a visual object")));
-- (instancetype)initWithCoder:(NSCoder *)aDecoder __attribute__((unavailable("please use 'initWithAppId:adType:delegate:' instead; the instance of this class should not be used as a visual object")));
-+ (instancetype)new __attribute__((unavailable("please use 'initWithAppId:adType:delegate:' instead")));
+- (null_unspecified instancetype)init __attribute__((unavailable("please use 'initWithAppId:adType:delegate:' instead")));
+- (null_unspecified instancetype)initWithFrame:(CGRect)frame __attribute__((unavailable("please use 'initWithAppId:adType:delegate:' instead; the instance of this class should not be used as a visual object")));
+- (null_unspecified instancetype)initWithCoder:(null_unspecified NSCoder *)aDecoder __attribute__((unavailable("please use 'initWithAppId:adType:delegate:' instead; the instance of this class should not be used as a visual object")));
++ (null_unspecified instancetype)new __attribute__((unavailable("please use 'initWithAppId:adType:delegate:' instead")));
 
 /**
  *  @brief Is used to control sound of a video ad in non-fullscreen mode.
  */
 @property (nonatomic) BOOL isMuted;
+
+/**
+ *  @brief Response data. Should be used only after 'InneractiveAdLoaded:' event has been received.
+ */
+@property (nullable, atomic, strong) IaNativeAdResponseModel *responseModel;
 
 /**
  *  @brief Plays a pre-loaded video ad.
@@ -78,7 +85,7 @@
  *
  *  @param adObject IaNativeAd instance.
  */
-- (void)layoutAdAssets:(IaNativeAd *)adObject;
+- (void)layoutAdAssets:(nonnull IaNativeAd *)adObject;
 
 @end
 
@@ -101,6 +108,42 @@
 
 @end
 
+#pragma mark - Native Ad Response Data Model
+
+@interface IaNativeAdResponseModel : NSObject
+
+/**
+ *  @brief Title;
+ */
+@property (nullable, nonatomic, strong) NSString *titleText; // IA supported / FB supported;
+
+/**
+ *  @brief Icon.
+ */
+@property (nullable, nonatomic, strong) NSString *iconImageURLString; // IA supported / FB supported;
+
+/**
+ *  @brief Video VAST 2.0 XML content as NSData.
+ */
+@property (nullable, nonatomic, strong) NSData *VASTXMLData; // IA supported / FB supported;
+
+/**
+ *  @brief Cover Image.
+ */
+@property (nullable, nonatomic, strong) NSString *largeImageURLString; // IA supported / FB supported;
+
+/**
+ *  @brief Description.
+ */
+@property (nullable, nonatomic, strong) NSString *descriptionString; // IA supported / FB supported;
+
+/**
+ *  @brief Call To Action text.
+ */
+@property (nullable, nonatomic, strong) NSString *CTAText; // IA supported / FB supported;
+
+@end
+
 #pragma mark - Native Ad Table Adapter
 
 /**
@@ -113,12 +156,12 @@
 /**
  *  @brief A UITableView instance, that is passed to and managed by IaNativeTableAdapter.
  */
-@property (nonatomic, strong, readonly) UITableView *table;
+@property (nonnull, nonatomic, strong, readonly) UITableView *table;
 
 /**
  *  @brief Class specification for ads cells.
  */
-@property (nonatomic, strong, readonly) Class adCellRegisteringClass;
+@property (nonnull, nonatomic, strong, readonly) Class adCellRegisteringClass;
 
 /**
  *  @brief A designated initializer.
@@ -131,7 +174,7 @@
  *
  *  @return                       IaNativeAdTableAdapter instance.
  */
-- (id)initWithNativeAd:(IaNativeAd *)nativeAd table:(UITableView *)table adCellRegisteringClass:(Class)adCellRegisteringClass;
+- (nonnull id)initWithNativeAd:(nonnull IaNativeAd *)nativeAd table:(nonnull UITableView *)table adCellRegisteringClass:(nonnull Class)adCellRegisteringClass;
 
 /**
  *  @brief Use this method before releasing the IaNativeAdTableAdapter instance.
@@ -142,8 +185,8 @@
 
 // The alloc, init and new methods should not be used to initialize and use the IaNativeAdTableAdapter class.
 // initWithNativeAd:table:adCellRegisteringClass: should be used instead.
-- (instancetype)init __attribute__((unavailable("init not available, call initWithNativeAd:table:adCellRegisteringClass: instead")));
-+ (instancetype)new __attribute__((unavailable("new not available, call initWithNativeAd:table:adCellRegisteringClass: instead")));
+- (null_unspecified instancetype)init __attribute__((unavailable("init not available, call initWithNativeAd:table:adCellRegisteringClass: instead")));
++ (null_unspecified instancetype)new __attribute__((unavailable("new not available, call initWithNativeAd:table:adCellRegisteringClass: instead")));
 
 @end
 
@@ -162,7 +205,7 @@
  * placer using, that table view, use this method rather than [UITableView setDataSource:].
  *  @param dataSource The new table view data source.
  */
-- (void)ia_setDataSource:(id<UITableViewDataSource>)dataSource;
+- (void)ia_setDataSource:(nullable id<UITableViewDataSource>)dataSource;
 
 /**
  *  @brief Returns the original data source of the table view.
@@ -172,7 +215,7 @@
  *
  *  @return The original table view data source.
  */
-- (id<UITableViewDataSource>)ia_dataSource;
+- (nullable id<UITableViewDataSource>)ia_dataSource;
 
 /**
  *  @brief Sets the table view's delegate.
@@ -181,7 +224,7 @@
  *
  *  @param delegate The new table view delegate.
  */
-- (void)ia_setDelegate:(id<UITableViewDelegate>)delegate;
+- (void)ia_setDelegate:(nullable id<UITableViewDelegate>)delegate;
 
 /**
  *  @brief Returns the original delegate of the table view.
@@ -191,7 +234,7 @@
  *
  *  @return The original table view delegate.
  */
-- (id<UITableViewDelegate>)ia_delegate;
+- (nullable id<UITableViewDelegate>)ia_delegate;
 
 /**
  *  @brief IA replacement for the original UITableView method. Use this method when working with table adapter.
@@ -231,60 +274,60 @@
 /**
  *  @brief IA replacement for the original UITableView method. Use this method when working with table adapter.
  */
-- (CGRect)ia_rectForRowAtIndexPath:(NSIndexPath *)indexPath;
+- (CGRect)ia_rectForRowAtIndexPath:(null_unspecified NSIndexPath *)indexPath;
 
 /**
  *  @brief IA replacement for the original UITableView method. Use this method when working with table adapter.
  *
  *  @return NSIndexPath instance. Returns nil if point is outside of any row in the table.
  */
-- (NSIndexPath *)ia_indexPathForRowAtPoint:(CGPoint)point;
+- (null_unspecified NSIndexPath *)ia_indexPathForRowAtPoint:(CGPoint)point;
 
 /**
  *  @brief IA replacement for the original UITableView method. Use this method when working with table adapter.
  *
  *  @return Nil if the cell is not visible.
  */
-- (NSIndexPath *)ia_indexPathForCell:(UITableViewCell *)cell;
+- (null_unspecified NSIndexPath *)ia_indexPathForCell:(null_unspecified UITableViewCell *)cell;
 
 /**
  *  @brief IA replacement for the original UITableView method. Use this method when working with table adapter.
  *
  *  @return Nil if the rect is not valid.
  */
-- (NSArray *)ia_indexPathsForRowsInRect:(CGRect)rect;
+- (null_unspecified NSArray *)ia_indexPathsForRowsInRect:(CGRect)rect;
 
 /**
  *  @brief IA replacement for the original UITableView method. Use this method when working with table adapter.
  *
  *  @return Nil if the cell is not visible or index path is out of range.
  */
-- (UITableViewCell *)ia_cellForRowAtIndexPath:(NSIndexPath *)indexPath;
+- (null_unspecified UITableViewCell *)ia_cellForRowAtIndexPath:(null_unspecified NSIndexPath *)indexPath;
 
 /**
  *  @brief IA replacement for the original UITableView method. Use this method when working with table adapter.
  */
-- (NSArray *)ia_visibleCells;
+- (null_unspecified NSArray *)ia_visibleCells;
 
 /**
  *  @brief IA replacement for the original UITableView method. Use this method when working with table adapter.
  */
-- (NSArray *)ia_indexPathsForVisibleRows;
+- (null_unspecified NSArray *)ia_indexPathsForVisibleRows;
 
 /**
  *  @brief IA replacement for the original UITableView method. Use this method when working with table adapter.
  */
-- (UITableViewHeaderFooterView *)ia_headerViewForSection:(NSInteger)section NS_AVAILABLE_IOS(6_0);
+- (null_unspecified UITableViewHeaderFooterView *)ia_headerViewForSection:(NSInteger)section NS_AVAILABLE_IOS(6_0);
 
 /**
  *  @brief IA replacement for the original UITableView method. Use this method when working with table adapter.
  */
-- (UITableViewHeaderFooterView *)ia_footerViewForSection:(NSInteger)section NS_AVAILABLE_IOS(6_0);
+- (null_unspecified UITableViewHeaderFooterView *)ia_footerViewForSection:(NSInteger)section NS_AVAILABLE_IOS(6_0);
 
 /**
  *  @brief IA replacement for the original UITableView method. Use this method when working with table adapter.
  */
-- (void)ia_scrollToRowAtIndexPath:(NSIndexPath *)indexPath atScrollPosition:(UITableViewScrollPosition)scrollPosition animated:(BOOL)animated;
+- (void)ia_scrollToRowAtIndexPath:(null_unspecified NSIndexPath *)indexPath atScrollPosition:(UITableViewScrollPosition)scrollPosition animated:(BOOL)animated;
 
 /**
  *  @brief IA replacement for the original UITableView method. Use this method when working with table adapter.
@@ -308,17 +351,17 @@
 /**
  *  @brief IA replacement for the original UITableView method. Use this method when working with table adapter.
  */
-- (void)ia_insertSections:(NSIndexSet *)sections withRowAnimation:(UITableViewRowAnimation)animation;
+- (void)ia_insertSections:(null_unspecified NSIndexSet *)sections withRowAnimation:(UITableViewRowAnimation)animation;
 
 /**
  *  @brief IA replacement for the original UITableView method. Use this method when working with table adapter.
  */
-- (void)ia_deleteSections:(NSIndexSet *)sections withRowAnimation:(UITableViewRowAnimation)animation;
+- (void)ia_deleteSections:(null_unspecified NSIndexSet *)sections withRowAnimation:(UITableViewRowAnimation)animation;
 
 /**
  *  @brief IA replacement for the original UITableView method. Use this method when working with table adapter.
  */
-- (void)ia_reloadSections:(NSIndexSet *)sections withRowAnimation:(UITableViewRowAnimation)animation NS_AVAILABLE_IOS(3_0);
+- (void)ia_reloadSections:(null_unspecified NSIndexSet *)sections withRowAnimation:(UITableViewRowAnimation)animation NS_AVAILABLE_IOS(3_0);
 
 /**
  *  @brief IA replacement for the original UITableView method. Use this method when working with table adapter.
@@ -328,55 +371,55 @@
 /**
  *  @brief IA replacement for the original UITableView method. Use this method when working with table adapter.
  */
-- (void)ia_insertRowsAtIndexPaths:(NSArray *)indexPaths withRowAnimation:(UITableViewRowAnimation)animation;
+- (void)ia_insertRowsAtIndexPaths:(null_unspecified NSArray *)indexPaths withRowAnimation:(UITableViewRowAnimation)animation;
 
 /**
  *  @brief IA replacement for the original UITableView method. Use this method when working with table adapter.
  */
-- (void)ia_deleteRowsAtIndexPaths:(NSArray *)indexPaths withRowAnimation:(UITableViewRowAnimation)animation;
+- (void)ia_deleteRowsAtIndexPaths:(null_unspecified NSArray *)indexPaths withRowAnimation:(UITableViewRowAnimation)animation;
 
 /**
  *  @brief IA replacement for the original UITableView method. Use this method when working with table adapter.
  */
-- (void)ia_reloadRowsAtIndexPaths:(NSArray *)indexPaths withRowAnimation:(UITableViewRowAnimation)animation NS_AVAILABLE_IOS(3_0);
+- (void)ia_reloadRowsAtIndexPaths:(null_unspecified NSArray *)indexPaths withRowAnimation:(UITableViewRowAnimation)animation NS_AVAILABLE_IOS(3_0);
 
 /**
  *  @brief IA replacement for the original UITableView method. Use this method when working with table adapter.
  */
-- (void)ia_moveRowAtIndexPath:(NSIndexPath *)indexPath toIndexPath:(NSIndexPath *)newIndexPath NS_AVAILABLE_IOS(5_0);
+- (void)ia_moveRowAtIndexPath:(null_unspecified NSIndexPath *)indexPath toIndexPath:(null_unspecified NSIndexPath *)newIndexPath NS_AVAILABLE_IOS(5_0);
 
 /**
  *  @brief IA replacement for the original UITableView method. Use this method when working with table adapter.
  *
  *  @return Nil or index path representing section and row of selection.
  */
-- (NSIndexPath *)ia_indexPathForSelectedRow;
+- (null_unspecified NSIndexPath *)ia_indexPathForSelectedRow;
 
 /**
  *  @brief IA replacement for the original UITableView method. Use this method when working with table adapter.
  *
  *  @return Nil or a set of index paths representing the sections and rows of the selection.
  */
-- (NSArray *)ia_indexPathsForSelectedRows NS_AVAILABLE_IOS(5_0);
+- (null_unspecified NSArray *)ia_indexPathsForSelectedRows NS_AVAILABLE_IOS(5_0);
 
 /**
  *  @brief IA replacement for the original UITableView method. Use this method when working with table adapter.
  */
-- (void)ia_selectRowAtIndexPath:(NSIndexPath *)indexPath animated:(BOOL)animated scrollPosition:(UITableViewScrollPosition)scrollPosition;
+- (void)ia_selectRowAtIndexPath:(null_unspecified NSIndexPath *)indexPath animated:(BOOL)animated scrollPosition:(UITableViewScrollPosition)scrollPosition;
 
 /**
  *  @brief IA replacement for the original UITableView method. Use this method when working with table adapter.
  */
-- (void)ia_deselectRowAtIndexPath:(NSIndexPath *)indexPath animated:(BOOL)animated;
+- (void)ia_deselectRowAtIndexPath:(null_unspecified NSIndexPath *)indexPath animated:(BOOL)animated;
 
 /**
  *  @brief IA replacement for the original UITableView method. Use this method when working with table adapter.
  */
-- (id)ia_dequeueReusableCellWithIdentifier:(NSString *)identifier forIndexPath:(NSIndexPath *)indexPath;
+- (null_unspecified id)ia_dequeueReusableCellWithIdentifier:(null_unspecified NSString *)identifier forIndexPath:(null_unspecified NSIndexPath *)indexPath;
 
 /**
  *  @brief IA replacement for the original UITableView method. Use this method when working with table adapter.
  */
-- (id)ia_dequeueReusableHeaderFooterViewWithIdentifier:(NSString *)identifier NS_AVAILABLE_IOS(6_0);
+- (null_unspecified id)ia_dequeueReusableHeaderFooterViewWithIdentifier:(null_unspecified NSString *)identifier NS_AVAILABLE_IOS(6_0);
 
 @end