+++ InneractiveAdSample/InneractiveAdSample/MainVCsAndAppDelegate/Story/TableViewExample/InneractiveNativeAdStoryTableAdapterViewController.h
@@ -6,7 +6,7 @@
 //  Copyright (c) 2014 Inneractive. All rights reserved.
 //
 
-#import <InneractiveAdSDK/InneractiveAdSDK.h>
+#import "InneractiveAdSDK.h"
 
 /**
  * Table Story example using table adapter